fix: check sig100 format for edge case
from time to time (on axel folder that we know, potentially others)
there is a problem on sim exp. this comes from sig100 not being correct,
it appears to be numbers instead of genes; which breaks everything.

added a check on app start to check if the h5 is formated fine, if not
it gets re-built